Fix broken link, and added course
Hovav Shacham now joined UT Austin, and the link in the book was broken You are absolutely right, his reading list is hard core, so I've updated it with a course he's teaching currently. I have also added a course taught by Nickolai Zeldovich, which is a great starting point for anyone trying to get a feel of security. You can add it, if you feel it's noteworthy.